TASK: Move RequestHandler.php
